Conflicts:
	src/controller/Loader.java

+++ src/model/constants/Interests.java
@@ -1,6 +1,6 @@
 package model.constants;
 /**
- * Classe qu'on peut utiliser pour retrouver facilement un interet donné. 
+ * Classe qu'on peut utiliser pour retrouver facilement un interet donnï¿½. 
  * Exemple: interestsBag.getInterests()[Interests.SPORTS] pour l'interet au sport
  * @author Martin
  *
@@ -13,21 +13,21 @@ public class Interests {
 	}
 
 	// Static constants
-	public static final byte SPORTS = 0;
-	public static final byte TVSPORTS = 1;
-	public static final byte EXERCISE = 2;
-	public static final byte DINING = 3;
-	public static final byte MUSEUMS = 4;
-	public static final byte ART = 5;
-	public static final byte HIKING = 6;
-	public static final byte GAMING = 7;
-	public static final byte CLUBBING = 8;
-	public static final byte READING = 9;
-	public static final byte TV = 10;
-	public static final byte THEATER = 11;
-	public static final byte MOVIES = 12;
-	public static final byte CONCERTS = 13;
-	public static final byte MUSIC = 14;
-	public static final byte SHOPPING = 15;
-	public static final byte YOGA = 16;
+	public static final int SPORTS = 0;
+	public static final int TVSPORTS = 1;
+	public static final int EXERCISE = 2;
+	public static final int DINING = 3;
+	public static final int MUSEUMS = 4;
+	public static final int ART = 5;
+	public static final int HIKING = 6;
+	public static final int GAMING = 7;
+	public static final int CLUBBING = 8;
+	public static final int READING = 9;
+	public static final int TV = 10;
+	public static final int THEATER = 11;
+	public static final int MOVIES = 12;
+	public static final int CONCERTS = 13;
+	public static final int MUSIC = 14;
+	public static final int SHOPPING = 15;
+	public static final int YOGA = 16;
 }
